+++ generators/testgen/src/testgen/io/testplans.py
@@ -1,13 +1,12 @@
 ##################################
-# testplans.py
+# io/testplans.py
 #
+# Read testplans for riscv-arch-test test generation.
 # jcarlin@hmc.edu 5 October 2025
 # SPDX-License-Identifier: Apache-2.0
 ##################################
 
-"""
-Read testplans for riscv-arch-test test generation.
-"""
+"""Read testplans for riscv-arch-test test generation."""
 
 import csv
 from dataclasses import dataclass
@@ -15,7 +14,7 @@ from pathlib import Path
 
 
 def get_extensions(testplan_dir: Path) -> list[str]:
-    """Get the list of extensions from the testplan directory"""
+    """Get the list of extensions from the testplan directory."""
     extensions: list[str] = []
     for testplan in testplan_dir.glob("*.csv"):
         extension = testplan.stem
@@ -27,9 +26,7 @@ def get_extensions(testplan_dir: Path) -> list[str]:
 
 @dataclass
 class TestPlanData:
-    """
-    Data structure for information on a single instruction parsed from a testplan.
-    """
+    """Data structure for information on a single instruction parsed from a testplan."""
 
     instr_type: str
     rv32: bool
@@ -38,7 +35,7 @@ class TestPlanData:
 
 
 def read_testplan(testplan_path: Path) -> dict[str, TestPlanData]:
-    """Read a testplan and return a dictionary of instructions and their associated data(type, coverpoints, etc.)"""
+    """Read a testplan and return a dictionary of instructions and their associated data (type, coverpoints, etc.)."""
     instructions: dict[str, TestPlanData] = {}
     with testplan_path.open() as csvfile:
         reader = csv.DictReader(csvfile)
